Change the Davis Shibboleth login so that it does not need to reset the iRODS user account passwords.
Documentation on this change can be found at

http://technical.bestgrid.org/index.php/DataFabric_Davis_enhancement_-_not_resetting_iRODS_password_on_Shibboleth_login

For reference, relevant exceprts from the initial version of the page are copied below.

== Davis Shibboleth Login changes ==

To overcome the need to reset the user password, the modifications introduced here instead use the iRODS protocol support for an admin user acting on behalf of another user - the equivalent of setting the '''<tt>clientUserName</tt>''' environment variable with i-Commands.

This feature is supported by the iRODS protocol, but not by the Jargon library.  The support has been added as a part of this enhancement and is pending commit to the main Jargon-classic repository (expected in the week of October 15th, 2012).

The Shibboleth login keeps the first three steps from the original workflow:

1. Davis would be only accessed via a Shibboleth-protected URL
2. Davis would receive the sharedToken and commonName attributes from the Shibboleth session
3. Davis would establish a session with the iRODS server as the rods admin user (using GSI authentication with the host certificate) 

But once Davis has the session as an administrator, it changes to:
4. Davis uses the administrator session to query the iRODS database for the username (based on the sharedToken value)
4.1. If the account is not found, it is created by invoking the '''createUser''' command and the search is repeated. 
5. Davis establishes a session with admin credentials used for authentication, but uses the username to act on behalf of the user.

As using a different client username does not work with GSI, Davis has to a username+password based login for the '''rods''' admin account.

=== Configuration options added ===

To support this new funcionality, this extension adds the following options to the Davis configuration file:

'''admin-creds-dir''': Set location of admin credentials.  When set, should point to a directory that has the login details for a rodsadmin account (typically the rods user).  The directory should contain two files:
* .irodsEnv at least setting irodsUserName - e.g.:<pre>irodsUserName 'rods'</pre>
* .irodsA with a single line containing the password of the rods user
* Default is no directory is set: <pre># admin-creds-dir=</pre>

'''shib-use-admin-login''': Enable this alternative login method for Shibboleth login that does not reset the password associated with the iRODS account but instead uses an admin account to establish the connection - but the connection runs with the privileges of the target user. 
* This option requires that the admin-creds-dir option is correctly configured to log in as a rodsadmin account (typically rods).
* Default is false: <pre>shib-use-admin-login=false</pre>

=== Implementation details ===

* Jargon is modified to extend the edu.sdsc.grid.io.irods.IRODSAccount API with support for setting a clientUserName
* webdavis.AuthorizationProcessor is modified to use the clientUserName login method when shib-use-admin-login is set to true
* webdavis.ShibUtil is modified:
** the code looking up an account by sharedToken is factored out into the <tt>lookupUsernameByST</tt> method
** the createUser command is only invoked if the user account is not found
* webdavis.DavisConfig is modified to parse the two new configuration options

=== Code access ===

The code has been committed to a feature-branch in the googlecode webDavis repository: http://webdavis.googlecode.com/svn/davis/branches/0.9.6.dev-shib-no-pwd-reset